+++ tests/helpers/utils.py
@@ -1,18 +1,18 @@
-"""This module contains utility code for unit tests.
+#  Copyright (c) 2023 https://reportportal.io .
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License
 
-Copyright (c) 2021 https://reportportal.io .
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
+"""This module contains utility code for unit tests."""
 
-https://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License
-"""
 import os
 import random
 import time
@@ -138,7 +138,7 @@ def attributes_to_tuples(attributes):
 # noinspection PyProtectedMember
 def run_tests_with_client(client, tests, args=None, variables=None):
     def test_func():
-        from reportportal_client._local import set_current
+        from reportportal_client import set_current
         set_current(client)
         return run_pytest_tests(tests, args, variables)