+++ tests/integration/test_attributes.py
@@ -1,4 +1,3 @@
-"""This module includes integration tests for different attribute reporting."""
 #  Copyright (c) 2022 https://reportportal.io .
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
@@ -11,6 +10,8 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
+
+"""This module includes integration tests for different attribute reporting."""
 
 from unittest import mock