+++ tests/integration/test_pass_failed_skipped.py
@@ -1,5 +1,4 @@
-"""This module includes integration tests for item statuses report."""
-#  Copyright (c) 2022 https://reportportal.io .
+#  Copyright (c) 2023 https://reportportal.io .
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 #  You may obtain a copy of the License at
@@ -11,6 +10,9 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
+
+"""This module includes integration tests for item statuses report."""
+
 import pytest
 from delayed_assert import expect, assert_expectations
 from unittest import mock