+++ tests/integration/test_threads_logs.py
@@ -1,3 +1,16 @@
+#  Copyright (c) 2023 https://reportportal.io .
+#  Licensed under the Apache License, Version 2.0 (the "License");
+#  you may not use this file except in compliance with the License.
+#  You may obtain a copy of the License at
+#
+#  https://www.apache.org/licenses/LICENSE-2.0
+#
+#  Unless required by applicable law or agreed to in writing, software
+#  distributed under the License is distributed on an "AS IS" BASIS,
+#  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+#  See the License for the specific language governing permissions and
+#  limitations under the License
+
 from unittest import mock
 
 from tests import REPORT_PORTAL_SERVICE
@@ -14,7 +27,7 @@ def test_rp_thread_logs_reporting(mock_client_init):
     mock_thread_client = mock_client.clone()
 
     def init_thread_client(*_, **__):
-        from reportportal_client._local import set_current
+        from reportportal_client import set_current
         set_current(mock_thread_client)
         return mock_thread_client