+++ tests/unit/conftest.py
@@ -1,6 +1,4 @@
-"""This module contains common Pytest fixtures and hooks for unit tests."""
-
-#  Copyright (c) 2021 https://reportportal.io .
+#  Copyright (c) 2023 https://reportportal.io .
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 #  You may obtain a copy of the License at
@@ -12,6 +10,8 @@
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
+
+"""This module contains common Pytest fixtures and hooks for unit tests."""
 
 import py
 from _pytest.config import Config
@@ -68,6 +68,8 @@ def mocked_config():
     mocked_config.option.rp_launch_timeout = -1
     mocked_config.option.rp_thread_logging = True
     mocked_config.option.rp_launch_uuid_print = 'False'
+    mocked_config.option.rp_launch_uuid_print_output = 'STDOUT'
+    mocked_config.option.rp_client_type = 'SYNC'
     return mocked_config