+++ tests/unit/test_plugin.py
@@ -1,17 +1,17 @@
-"""This module includes unit tests for the plugin."""
-
-#  Copyright (c) 2021 http://reportportal.io .
+#  Copyright (c) 2023 https://reportportal.io .
 #  Licensed under the Apache License, Version 2.0 (the "License");
 #  you may not use this file except in compliance with the License.
 #  You may obtain a copy of the License at
 #
-#  http://www.apache.org/licenses/LICENSE-2.0
+#  https://www.apache.org/licenses/LICENSE-2.0
 #
 #  Unless required by applicable law or agreed to in writing, software
 #  distributed under the License is distributed on an "AS IS" BASIS,
 #  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 #  See the License for the specific language governing permissions and
 #  limitations under the License
+
+"""This module includes unit tests for the plugin."""
 
 from _pytest.config.argparsing import Parser
 import pytest
@@ -366,7 +366,10 @@ def test_pytest_addoption_adds_correct_ini_file_arguments():
         'retries',
         'rp_api_retries',
         'rp_skip_connection_test',
-        'rp_launch_timeout'
+        'rp_launch_timeout',
+        'rp_client_type',
+        'rp_connect_timeout',
+        'rp_read_timeout'
     )
     mock_parser = mock.MagicMock(spec=Parser)